backend/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
Spire.Presentation==8.9.0
